feat(admin) : warning 컴포넌트 component 폴더로 이동 (#125)
